move things that belong into window into window
