+++ src/vs/code/electron-main/window.ts
@@ -9,6 +9,7 @@ import * as path from 'path';
 import * as objects from 'vs/base/common/objects';
 import { stopProfiling } from 'vs/base/node/profiler';
 import nls = require('vs/nls');
+import URI from "vs/base/common/uri";
 import { IStorageService } from 'vs/platform/storage/node/storage';
 import { shell, screen, BrowserWindow, systemPreferences, app } from 'electron';
 import { TPromise, TValueCallback } from 'vs/base/common/winjs.base';
@@ -110,10 +111,6 @@ export class CodeWindow implements ICodeWindow {
 		// respect configured menu bar visibility
 		this.onConfigurationUpdated();
 
-		// TODO@joao: hook this up to some initialization routine this causes a race between setting the headers and doing
-		// a request that needs them. chances are low
-		this.setCommonHTTPHeaders();
-
 		// Eventing
 		this.registerListeners();
 	}
@@ -201,20 +198,6 @@ export class CodeWindow implements ICodeWindow {
 		}
 
 		this._lastFocusTime = Date.now(); // since we show directly, we need to set the last focus time too
-	}
-
-	private setCommonHTTPHeaders(): void {
-		getCommonHTTPHeaders().done(headers => {
-			if (!this._win) {
-				return;
-			}
-
-			const urls = ['https://marketplace.visualstudio.com/*', 'https://*.vsassets.io/*'];
-
-			this._win.webContents.session.webRequest.onBeforeSendHeaders({ urls }, (details, cb) => {
-				cb({ cancel: false, requestHeaders: objects.assign(details.requestHeaders, headers) });
-			});
-		});
 	}
 
 	public hasHiddenTitleBarStyle(): boolean {
@@ -314,6 +297,42 @@ export class CodeWindow implements ICodeWindow {
 	}
 
 	private registerListeners(): void {
+
+		// Set common HTTP headers
+		// TODO@joao: hook this up to some initialization routine this causes a race between setting the headers and doing
+		// a request that needs them. chances are low
+		getCommonHTTPHeaders().done(headers => {
+			if (!this._win) {
+				return;
+			}
+
+			const urls = ['https://marketplace.visualstudio.com/*', 'https://*.vsassets.io/*'];
+
+			this._win.webContents.session.webRequest.onBeforeSendHeaders({ urls }, (details, cb) => {
+				cb({ cancel: false, requestHeaders: objects.assign(details.requestHeaders, headers) });
+			});
+		});
+
+		// Prevent loading of svgs
+		this._win.webContents.session.webRequest.onBeforeRequest((details, callback) => {
+			if (details.url.indexOf('.svg') > 0) {
+				const uri = URI.parse(details.url);
+				if (uri && !uri.scheme.match(/file/i) && (uri.path as any).endsWith('.svg')) {
+					return callback({ cancel: true });
+				}
+			}
+
+			return callback({});
+		});
+
+		this._win.webContents.session.webRequest.onHeadersReceived((details, callback) => {
+			const contentType: string[] = (details.responseHeaders['content-type'] || details.responseHeaders['Content-Type']) as any;
+			if (contentType && Array.isArray(contentType) && contentType.some(x => x.toLowerCase().indexOf('image/svg') >= 0)) {
+				return callback({ cancel: true });
+			}
+
+			return callback({ cancel: false, responseHeaders: details.responseHeaders });
+		});
 
 		// Remember that we loaded
 		this._win.webContents.on('did-finish-load', () => {
