+++ src/vs/code/electron-main/windows.ts
@@ -10,7 +10,6 @@ import * as fs from 'original-fs';
 import * as nls from 'vs/nls';
 import * as arrays from 'vs/base/common/arrays';
 import { assign, mixin } from 'vs/base/common/objects';
-import URI from 'vs/base/common/uri';
 import { IBackupMainService } from 'vs/platform/backup/common/backup';
 import { IEnvironmentService, ParsedArgs } from 'vs/platform/environment/common/environment';
 import { IStorageService } from 'vs/platform/storage/node/storage';
@@ -835,25 +834,6 @@ export class WindowsManager implements IWindowsMainService {
 			codeWindow.win.webContents.on('crashed', () => this.onWindowError(codeWindow, WindowError.CRASHED));
 			codeWindow.win.on('unresponsive', () => this.onWindowError(codeWindow, WindowError.UNRESPONSIVE));
 			codeWindow.win.on('closed', () => this.onWindowClosed(codeWindow));
-
-			// Prevent loading on svgs in main renderer
-			codeWindow.win.webContents.session.webRequest.onBeforeRequest((details, callback) => {
-				if (details.url.indexOf('.svg') > 0) {
-					const uri = URI.parse(details.url);
-					if (uri && !uri.scheme.match(/file/i) && (uri.path as any).endsWith('.svg')) {
-						return callback({ cancel: true });
-					}
-				}
-				return callback({});
-			});
-
-			codeWindow.win.webContents.session.webRequest.onHeadersReceived((details, callback) => {
-				const contentType: string[] = (details.responseHeaders['content-type'] || details.responseHeaders['Content-Type']) as any;
-				if (contentType && Array.isArray(contentType) && contentType.some(x => x.toLowerCase().indexOf('image/svg') >= 0)) {
-					return callback({ cancel: true });
-				}
-				return callback({ cancel: false, responseHeaders: details.responseHeaders });
-			});
 
 			// Lifecycle
 			this.lifecycleService.registerWindow(codeWindow);
